error correction reported by toss_007

+++ scripts.php
@@ -15,20 +15,20 @@ return [
 			$util->createTable('@adsensekit_ads', function ($table) {
 				$table->addColumn('id', 'integer', ['unsigned' => true, 'length' => 10, 'autoincrement' => true]);
 				$table->addColumn('status', 'smallint');
-        $table->addColumn('priority', 'smallint', ['default' => 0]);
+				$table->addColumn('priority', 'smallint', ['default' => 0]);
 				$table->addColumn('title', 'string', ['length' => 255]);
 				$table->addColumn('slug', 'string', ['length' => 255]);
-				$table->addColumn('adtype', 'tinyint', ['default' => 0]);
+				$table->addColumn('adtype', 'smallint', ['default' => 0]);
 				$table->addColumn('adsrc_img', 'string', ['length' => 255]);
 				$table->addColumn('adurl_img', 'string', ['length' => 255]);
 				$table->addColumn('adformat_adsense', 'string', ['length' => 255]);
-        $table->addColumn('adlayoutkey_adsense', 'string', ['length' => 255]);
-        $table->addColumn('adclient_adsense', 'string', ['length' => 255]);
-        $table->addColumn('adslot_adsense', 'string', ['length' => 255]);
+				$table->addColumn('adlayoutkey_adsense', 'string', ['length' => 255]);
+				$table->addColumn('adclient_adsense', 'string', ['length' => 255]);
+				$table->addColumn('adslot_adsense', 'string', ['length' => 255]);
 				$table->addColumn('date_start', 'datetime', ['notnull' => false]);
 				$table->addColumn('date_end', 'datetime', ['notnull' => false]);
 				$table->setPrimaryKey(['id']);
-        $table->addUniqueIndex(['slug'], '@ADSENSEKIT_SLUG');
+				$table->addUniqueIndex(['slug'], '@ADSENSEKIT_SLUG');
 			});
 		}
